Update AdderWS waybar conf
(mostly) fixed a bug that closes waybar when you scroll on moduals that dont usally allow scrolling. like network or the CPU gauge. it will send a notification instead. But if you scroll in that tiny space between the brightness CTL and the network monitor it will close. every area besides that is fine though.